server: dockerfile: fix broken ENTRYPOINT from generated project
This cost me an entire afternoon. I don't know whose bright idea it
was to just slap all arguments into the ENTRYPOINT, but here we are.

ENTRYPOINT now just runs the `server` executable, while COMMAND is
used to pass arguments / flags to it. This is useful for e.g. passing
`--apply-migrations` for deployments.

Signed-off-by: Max R. Carrara <max@aequito.sh>