server: docker: fix traefik not being able to discover server
tl;dr: Since the server is in both the postgres instance's isolated
"backend" network and the public-facing "docker-net", traefik
unfortunately happened to decide to use the "backend" network for its
proxying purposes. This led to the server not being reachable at all.

This commit fixes that by explicitly telling traefik which network to
use.

Signed-off-by: Max R. Carrara <max@aequito.sh>